requirements.txt
-e git+https://github.com/acados/acados.git@f9075e7f705a3a1df94bb6868db14342fde7b90b#egg=acados_template&subdirectory=interfaces/acados_template
actionlib==1.13.2
angles==1.9.13
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bondpy==1.8.6
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
casadi==3.5.1
catkin==0.8.10
catkin-pkg==0.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.1
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
docopt==0.6.2
docutils==0.19
dynamic-reconfigure==1.7.3
empy==3.3.4
entrypoints==0.4
evdev==1.6.0
fastjsonschema==2.16.2
future-fstrings==1.2.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
idna==3.4
image-geometry==1.16.1
imageio==2.22.1
importlib-metadata==5.0.0
importlib-resources==5.10.0
interactive-markers==1.12.0
ipykernel==6.16.0
ipython==7.34.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
laser_geometry==1.6.7
MarkupSafe==2.1.1
matplotlib==3.3.1
matplotlib-inline==0.1.6
message-filters==1.15.14
mistune==2.0.4
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.4.12
numpy==1.18.1
opencv-python==4.6.0.66
osrf-pycommon==2.0.2
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.9.7
Pygments==2.13.0
pynput==1.7.6
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-qt-binding==0.4.4
python-xlib==0.31
pytz==2022.4
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
quadprog==0.1.7
requests==2.28.1
resource_retriever==1.12.7
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospy==1.15.14
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-moveit==0.5.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.11
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.9
rqt_logger_level==0.4.11
rqt_msg==0.4.10
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.11
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_robot_steering==0.5.12
rqt_runtime_monitor==0.5.9
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_tf_tree==0.6.3
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.19
scikit-image==0.19.3
scikit-learn==0.23.1
scipy==1.7.3
Send2Trash==1.8.0
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
soupsieve==2.3.2.post1
terminado==0.16.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
topic-tools==1.15.14
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
trajectory-planning-helpers==0.76
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
xacro==1.14.13
yarg==0.1.9
zipp==3.9.0
